Create README.md
I feel honoured to be able to present my work to a group of international researchers across the globe and am thankful for this stage. This work is done under the supervision of Dr. Diana Besliu Ionescu from the Astronomical Institute of the Romanian Academy. This is at its early development phase, where we have been working out the period of positive activity of solar quakes over a period of time(1 month, based on our test)